[AIRFLOW-5656] Rename provider to providers module
