+++ airflow/providers/google/marketing_platform/hooks/campaign_manager.py
