+++ docs/conf.py
@@ -208,10 +208,10 @@ exclude_patterns = [
     '_api/airflow/gcp/index.rst',
     '_api/airflow/gcp/example_dags',
     '_api/airflow/gcp/utils',
-    '_api/airflow/provider/index.rst',
-    '_api/airflow/provider/google/index.rst',
-    '_api/airflow/provider/google/marketing_platform/index.rst',
-    '_api/airflow/provider/google/marketing_platform/example_dags',
+    '_api/airflow/providers/index.rst',
+    '_api/airflow/providers/google/index.rst',
+    '_api/airflow/providers/google/marketing_platform/index.rst',
+    '_api/airflow/providers/google/marketing_platform/example_dags',
     'autoapi_templates',
     'howto/operator/gcp/_partials',
 ]
