+++ tests/providers/google/marketing_platform/hooks/__init__.py
