+++ tests/providers/google/marketing_platform/operators/__init__.py
