+++ tests/providers/google/marketing_platform/operators/test_campaign_manager_system.py
@@ -20,7 +20,7 @@ import unittest
 
 from tests.gcp.utils.base_gcp_system_test_case import SKIP_TEST_WARNING, TestDagGcpSystem
 from tests.gcp.utils.gcp_authenticator import GOOGLE_CAMPAIGN_MANAGER_KEY
-from tests.provider.google.marketing_platform.operators.test_campaign_manager_system_helper import (
+from tests.providers.google.marketing_platform.operators.test_campaign_manager_system_helper import (
     GoogleCampaignManagerTestHelper,
 )
