+++ tests/providers/google/marketing_platform/operators/test_campaign_manager_system_helper.py
