+++ tests/providers/google/marketing_platform/sensors/__init__.py
